refactor(markdown-validator): rename package to info.jab.mv
