+++ AGENTS.md
@@ -63,7 +63,7 @@ This project uses **OpenSpec** for structured change management and planning:
 ./mvnw clean verify -pl skills-generator
 
 # Validate Markdown documentation (requires JBang)
-jbang markdown-validator/src/main/java/info/jab/markdownvalidator/MarkdownValidator.java .
+jbang markdown-validator/src/main/java/info/jab/mv/MarkdownValidator.java .
 
 # Generate and copy skills for local agent use without touching skills/
 ./mvnw clean install -pl skills-generator
@@ -166,6 +166,6 @@ pre-commit run conventional-pre-commit --hook-stage commit-msg --commit-msg-file
 
 ## Boundaries
 
-- ✅ **Always do:** Choose validation based on the files changed. For Markdown-only documentation changes outside `site-generator/`, run `jbang markdown-validator/src/main/java/info/jab/markdownvalidator/MarkdownValidator.java .` when JBang is available and inspect changed local links; do **not** run `./mvnw clean verify`. For Java, Maven, XML generator, or other build-affecting changes, run the relevant module verification or `./mvnw clean verify` before promoting. Edit XML in `skills-generator/src/main/resources/` (`skill-references/`, `skills/`) to change rules and skills, validate edited XML with `xmllint --noout <path-to-edited-file.xml>`, and follow PML Schema: [https://jabrena.github.io/pml/schemas/0.7.0/pml.xsd](https://jabrena.github.io/pml/schemas/0.7.0/pml.xsd). For local skill regeneration, use `./mvnw clean install -pl skills-generator` and test the generated output from `.agents/skills`; do not refresh `skills/` unless preparing an intentional release. For release skill changes, run `./mvnw clean install -pl skills-generator -P release`, then validate `skills/` with `npx skill-check@latest skills --no-security-scan --format github` and `skill-scanner scan-all ./skills --recursive --use-behavioral --policy strict --fail-on-severity high` when the scanner is available. For complex changes, create OpenSpec proposals first. When website sources change, regenerate and review `docs`, then commit source and output together. When you change `README.md`, update the translated READMEs (`README_ES.md`, `README_ZH.md`) in the same change. When you change an English `documentation/guides/GETTING-STARTED-*.md` file, update the matching `_ES.md` and `_ZH.md` files in the same change when they exist.
+- ✅ **Always do:** Choose validation based on the files changed. For Markdown-only documentation changes outside `site-generator/`, run `jbang markdown-validator/src/main/java/info/jab/mv/MarkdownValidator.java .` when JBang is available and inspect changed local links; do **not** run `./mvnw clean verify`. For Java, Maven, XML generator, or other build-affecting changes, run the relevant module verification or `./mvnw clean verify` before promoting. Edit XML in `skills-generator/src/main/resources/` (`skill-references/`, `skills/`) to change rules and skills, validate edited XML with `xmllint --noout <path-to-edited-file.xml>`, and follow PML Schema: [https://jabrena.github.io/pml/schemas/0.7.0/pml.xsd](https://jabrena.github.io/pml/schemas/0.7.0/pml.xsd). For local skill regeneration, use `./mvnw clean install -pl skills-generator` and test the generated output from `.agents/skills`; do not refresh `skills/` unless preparing an intentional release. For release skill changes, run `./mvnw clean install -pl skills-generator -P release`, then validate `skills/` with `npx skill-check@latest skills --no-security-scan --format github` and `skill-scanner scan-all ./skills --recursive --use-behavioral --policy strict --fail-on-severity high` when the scanner is available. For complex changes, create OpenSpec proposals first. When website sources change, regenerate and review `docs`, then commit source and output together. When you change `README.md`, update the translated READMEs (`README_ES.md`, `README_ZH.md`) in the same change. When you change an English `documentation/guides/GETTING-STARTED-*.md` file, update the matching `_ES.md` and `_ZH.md` files in the same change when they exist.
 - ⚠️ **Ask first:** Adding new XML rule files, modifying the XSLT stylesheet, changing site templates, architectural changes (use OpenSpec for planning)
 - 🚫 **Never do:** Edit `.cursor/rules/` or `docs/` directly, run `./mvnw clean verify` for Markdown-only documentation changes, commit secrets, skip applicable validation before promoting, bypass OpenSpec for major changes
