--- markdown-validator/src/main/java/info/jab/markdownvalidator/application/port/MarkdownFileFinder.java
@@ -1,9 +1,0 @@
-package info.jab.markdownvalidator.application.port;
-
-import java.io.IOException;
-import java.nio.file.Path;
-import java.util.List;
-
-public interface MarkdownFileFinder {
-    List<Path> findMarkdownFiles(Path root, List<String> targetDirectories) throws IOException;
-}
