--- markdown-validator/src/main/java/info/jab/markdownvalidator/application/port/RemoteLinkResponse.java
@@ -1,3 +1,0 @@
-package info.jab.markdownvalidator.application.port;
-
-public record RemoteLinkResponse(int statusCode) {}
