--- markdown-validator/src/main/java/info/jab/markdownvalidator/domain/ParseSummary.java
@@ -1,3 +1,0 @@
-package info.jab.markdownvalidator.domain;
-
-public record ParseSummary(String fileName, int markdownCharacters, int htmlCharacters) {}
