--- markdown-validator/src/main/java/info/jab/markdownvalidator/domain/ValidationError.java
@@ -1,5 +1,0 @@
-package info.jab.markdownvalidator.domain;
-
-import java.nio.file.Path;
-
-public record ValidationError(Path file, int lineNumber, String message) {}
