+++ markdown-validator/src/main/java/info/jab/mv/application/MarkdownValidationService.java
@@ -1,0 +1,34 @@
+package info.jab.mv.application;
+
+import info.jab.mv.application.port.MarkdownFileFinder;
+import info.jab.mv.domain.FileValidationResult;
+import info.jab.mv.domain.ValidationReport;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.List;
+
+public final class MarkdownValidationService {
+
+    private final MarkdownFileFinder fileFinder;
+    private final StructuredValidationRunner validationRunner;
+
+    public MarkdownValidationService(MarkdownFileFinder fileFinder, StructuredValidationRunner validationRunner) {
+        this.fileFinder = fileFinder;
+        this.validationRunner = validationRunner;
+    }
+
+    public ValidationReport validate(Path root, List<String> targetDirectories) throws IOException {
+        if (!Files.exists(root)) {
+            return ValidationReport.rootMissing(root);
+        }
+
+        List<Path> markdownFiles = fileFinder.findMarkdownFiles(root, targetDirectories);
+        if (markdownFiles.isEmpty()) {
+            return ValidationReport.empty(root);
+        }
+
+        List<FileValidationResult> results = validationRunner.validate(markdownFiles);
+        return ValidationReport.completed(root, markdownFiles.size(), results);
+    }
+}
