+++ markdown-validator/src/main/java/info/jab/mv/application/port/MarkdownFileFinder.java
@@ -1,0 +1,9 @@
+package info.jab.mv.application.port;
+
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.List;
+
+public interface MarkdownFileFinder {
+    List<Path> findMarkdownFiles(Path root, List<String> targetDirectories) throws IOException;
+}
