+++ markdown-validator/src/main/java/info/jab/mv/application/port/RemoteLinkRequester.java
@@ -1,0 +1,9 @@
+package info.jab.mv.application.port;
+
+import java.io.IOException;
+import java.net.URI;
+import java.time.Duration;
+
+public interface RemoteLinkRequester {
+    RemoteLinkResponse request(URI uri, String method, Duration timeout) throws IOException, InterruptedException;
+}
