+++ markdown-validator/src/main/java/info/jab/mv/application/port/RemoteLinkResponse.java
@@ -1,0 +1,3 @@
+package info.jab.mv.application.port;
+
+public record RemoteLinkResponse(int statusCode) {}
