+++ markdown-validator/src/main/java/info/jab/mv/domain/ParseSummary.java
@@ -1,0 +1,3 @@
+package info.jab.mv.domain;
+
+public record ParseSummary(String fileName, int markdownCharacters, int htmlCharacters) {}
