+++ markdown-validator/src/main/java/info/jab/mv/domain/ValidationError.java
@@ -1,0 +1,5 @@
+package info.jab.mv.domain;
+
+import java.nio.file.Path;
+
+public record ValidationError(Path file, int lineNumber, String message) {}
