--- markdown-validator/src/test/java/info/jab/markdownvalidator/adapter/in/cli/MarkdownValidatorTest.java
@@ -1,102 +1,0 @@
-package info.jab.markdownvalidator.adapter.in.cli;
-
-import info.jab.markdownvalidator.adapter.out.filesystem.FileSystemMarkdownFileFinder;
-import info.jab.markdownvalidator.application.MarkdownDocumentValidator;
-import info.jab.markdownvalidator.application.MarkdownValidationService;
-import info.jab.markdownvalidator.application.RemoteLinkValidator;
-import info.jab.markdownvalidator.application.StructuredValidationRunner;
-import info.jab.markdownvalidator.application.port.RemoteLinkRequester;
-import info.jab.markdownvalidator.application.port.RemoteLinkResponse;
-import java.io.IOException;
-import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.time.Duration;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-import picocli.CommandLine;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class MarkdownValidatorTest {
-
-    @TempDir
-    Path tempDir;
-
-    @Test
-    void commandLine_preservesCurrentOptionsAndDefaults() {
-        MarkdownValidatorCommand validator = validator(new FixedRequester(200));
-
-        new CommandLine(validator).parseArgs("--directories", "docs,skills", tempDir.toString());
-
-        assertThat(validator.targetDirectories).containsExactly("docs", "skills");
-        assertThat(validator.rootDir).isEqualTo(tempDir.toString());
-
-        MarkdownValidatorCommand defaultValidator = validator(new FixedRequester(200));
-
-        new CommandLine(defaultValidator).parseArgs();
-
-        assertThat(defaultValidator.targetDirectories).isEqualTo(MarkdownValidatorCommand.DEFAULT_TARGET_DIRECTORIES);
-        assertThat(defaultValidator.rootDir).isEqualTo(".");
-    }
-
-    @Test
-    void call_returnsFailureExitCodeWhenRootDoesNotExist() {
-        MarkdownValidatorCommand validator = validator(new FixedRequester(200));
-        String missingRoot = tempDir.resolve("missing").toString();
-
-        int exitCode = new CommandLine(validator).execute(missingRoot);
-
-        assertThat(exitCode).isEqualTo(1);
-    }
-
-    @Test
-    void call_succeedsWhenTargetDirectoriesAreMissing() {
-        MarkdownValidatorCommand validator = validator(new FixedRequester(200));
-        validator.targetDirectories = java.util.List.of("missing");
-
-        int exitCode = new CommandLine(validator).execute(tempDir.toString());
-
-        assertThat(exitCode).isZero();
-    }
-
-    @Test
-    void call_parsesMarkdownRendersHtmlAndReturnsSuccess() throws IOException {
-        Path docs = Files.createDirectories(tempDir.resolve("docs"));
-        Files.writeString(docs.resolve("guide.md"), "# Title%n%n[OK](https://example.test/ok)%n".formatted());
-        MarkdownValidatorCommand validator = validator(new FixedRequester(200));
-        validator.targetDirectories = java.util.List.of("docs");
-
-        int exitCode = new CommandLine(validator).execute(tempDir.toString());
-
-        assertThat(exitCode).isZero();
-    }
-
-    @Test
-    void call_preservesRemoteFailureExitCodeAndReportShape() throws IOException {
-        Path docs = Files.createDirectories(tempDir.resolve("docs"));
-        Files.writeString(docs.resolve("broken.md"), "[bad](https://example.test/missing)%n".formatted());
-        MarkdownValidatorCommand validator = validator(new FixedRequester(404));
-        validator.targetDirectories = java.util.List.of("docs");
-
-        int exitCode = new CommandLine(validator).execute(tempDir.toString());
-
-        assertThat(exitCode).isEqualTo(1);
-    }
-
-    private static MarkdownValidatorCommand validator(RemoteLinkRequester requester) {
-        RemoteLinkValidator remoteLinkValidator = new RemoteLinkValidator(requester, Duration.ofSeconds(10));
-        MarkdownDocumentValidator documentValidator = new MarkdownDocumentValidator(remoteLinkValidator);
-        MarkdownValidationService validationService =
-                new MarkdownValidationService(new FileSystemMarkdownFileFinder(), new StructuredValidationRunner(documentValidator));
-        return new MarkdownValidatorCommand(validationService);
-    }
-
-    private record FixedRequester(int statusCode) implements RemoteLinkRequester {
-
-        @Override
-        public RemoteLinkResponse request(URI uri, String method, Duration timeout) {
-            return new RemoteLinkResponse(statusCode);
-        }
-    }
-}
